Solved multiple unnecessary images
Filenames now gets encrypted by filename + filesize instead of filename
+ date(YmdHis). This should solve the problem if the user dont upload
identical name + size files. Issue #220

+++ PlayerEditor/imagerecorder2/upload.php
@@ -28,18 +28,21 @@
 			$filename = basename($file['name']);
 			$fileext = substr(strrchr($filename,'.'),1);
 			// Encrypt file name to ensure duplicate file names can co-exist (16 char names)
-			$encrypted_filename = "i".substr(md5($filename.date("YmdHis")),0,15);
+			$encrypted_filename = "i".substr(md5($filename.$file['size']),0,15);
 			
 			// Put it all together
 			$finalPath = $uploaddir."/".$encrypted_filename.".".$fileext;
 			
-			// Upload the file.
-			if(move_uploaded_file($file["tmp_name"], $finalPath)) {
-				$files[] = $uploaddir.$file["name"];
-			}
-			else {
-				$error = true;
-				$data = array("ERROR" => "Couldn't execute move_uploaded_file.");
+			// If file doesn't already exist upload it.
+			if(!file_exists($finalPath)) {
+				// Upload the file.
+				if(move_uploaded_file($file["tmp_name"], $finalPath)) {
+					$files[] = $uploaddir.$file["name"];
+				}
+				else {
+					$error = true;
+					$data = array("ERROR" => "Couldn't execute move_uploaded_file.");
+				}
 			}
 		}		
 		if(!$error) {
